fix(redux): fix posts' reducer for lastVisited array - add significant constraints
Before this commit there were no checks for post.post_id-s we were adding
in the lastVisited array. So, there was possibility that it may contain
repetitions and it can increase as much as we want. But, we actually want
save only 10.

From this commit there are 2 constarints for setPostsLastVisited case:
	1) if it is already in the array, then we return untouched state
	2) if lastVisited array[] contains 10 elements, we take first 9
		and unshift() new one in the start of it
		(so, we pop the oldest one)

modified: src/store/slices/Posts/Posts.slice.js